+++ src/components/Shop/Shop.js
@@ -1,222 +1,21 @@
-import React, { useState, useEffect, useContext } from "react";
-import { OpenCartContext } from "../../contexts/OpenCartContext";
-import { useQuery, useMutation } from "@apollo/react-hooks";
+import React, { useContext } from "react";
+import { CartContext } from "../../contexts/CartContext";
 import Product from "./Product";
-import Cart from "./Cart";
 import CustomerAuthWithMutation from "./CustomerAuth";
-import gql from "graphql-tag";
-import {
-  useCheckoutEffect,
-  createCheckout,
-  checkoutLineItemsAdd,
-  checkoutLineItemsUpdate,
-  checkoutLineItemsRemove,
-  checkoutCustomerAssociate,
-} from "../../queries/checkout";
 
-// TODO: All Cart props have to be Contexts so we can access Cart from About page (OR About page has no route?)
-
-const query = gql`
-  query query {
-    shop {
-      name
-      description
-      products(first: 20) {
-        pageInfo {
-          hasNextPage
-          hasPreviousPage
-        }
-        edges {
-          node {
-            id
-            title
-            options {
-              id
-              name
-              values
-            }
-            variants(first: 250) {
-              pageInfo {
-                hasNextPage
-                hasPreviousPage
-              }
-              edges {
-                node {
-                  id
-                  title
-                  selectedOptions {
-                    name
-                    value
-                  }
-                  image {
-                    src
-                  }
-                  price
-                }
-              }
-            }
-            images(first: 250) {
-              pageInfo {
-                hasNextPage
-                hasPreviousPage
-              }
-              edges {
-                node {
-                  src
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-`;
-
-function Shop(props) {
-  const { isCartOpen, setCartOpen } = useContext(OpenCartContext);
-
-  const [isNewCustomer, setNewCustomer] = useState(false);
-  const [isCustomerAuthOpen, setCustomerAuthOpen] = useState(false);
-  const [showAccountVerificationMessage, setAccountVerificationMessage] =
-    useState(false);
-  const [checkout, setCheckout] = useState({ lineItems: { edges: [] } });
-
-  const [customerAccessToken, setCustomerAccessToken] = useState(null);
-
-  const [
-    createCheckoutMutation,
-    {
-      data: createCheckoutData,
-      loading: createCheckoutLoading,
-      error: createCheckoutError,
-    },
-  ] = useMutation(createCheckout);
-
-  const [
-    lineItemAddMutation,
-    {
-      data: lineItemAddData,
-      loading: lineItemAddLoading,
-      error: lineItemAddError,
-    },
-  ] = useMutation(checkoutLineItemsAdd);
-
-  const [
-    lineItemUpdateMutation,
-    {
-      data: lineItemUpdateData,
-      loading: lineItemUpdateLoading,
-      error: lineItemUpdateError,
-    },
-  ] = useMutation(checkoutLineItemsUpdate);
-
-  const [
-    lineItemRemoveMutation,
-    {
-      data: lineItemRemoveData,
-      loading: lineItemRemoveLoading,
-      error: lineItemRemoveError,
-    },
-  ] = useMutation(checkoutLineItemsRemove);
-
-  const [
-    customerAssociateMutation,
-    {
-      data: customerAssociateData,
-      loading: customerAssociateLoading,
-      error: customerAssociateError,
-    },
-  ] = useMutation(checkoutCustomerAssociate);
-
+function Shop() {
   const {
-    loading: shopLoading,
-    error: shopError,
-    data: shopData,
-  } = useQuery(query);
-
-  useEffect(() => {
-    const variables = { input: {} };
-    createCheckoutMutation({ variables }).then(
-      (res) => {
-        console.log(res);
-      },
-      (err) => {
-        console.log("create checkout error", err);
-      }
-    );
-  }, []);
-
-  useCheckoutEffect(createCheckoutData, "checkoutCreate", setCheckout);
-  useCheckoutEffect(lineItemAddData, "checkoutLineItemsAdd", setCheckout);
-  useCheckoutEffect(lineItemUpdateData, "checkoutLineItemsUpdate", setCheckout);
-  useCheckoutEffect(lineItemRemoveData, "checkoutLineItemsRemove", setCheckout);
-  useCheckoutEffect(
-    customerAssociateData,
-    "checkoutCustomerAssociate",
-    setCheckout
-  );
-
-  const handleCartClose = () => {
-    setCartOpen(false);
-  };
-
-  const openCustomerAuth = (event) => {
-    if (event.target.getAttribute("data-customer-type") === "new-customer") {
-      setNewCustomer(true);
-      setCustomerAuthOpen(true);
-    } else {
-      setNewCustomer(false);
-      setCustomerAuthOpen(true);
-    }
-  };
-
-  const accountVerificationMessage = () => {
-    setAccountVerificationMessage(true);
-    setTimeout(() => {
-      setAccountVerificationMessage(false);
-    }, 5000);
-  };
-
-  const closeCustomerAuth = () => {
-    setCustomerAuthOpen(false);
-  };
-
-  const addVariantToCart = (variantId, quantity) => {
-    const variables = {
-      checkoutId: checkout.id,
-      lineItems: [{ variantId, quantity: parseInt(quantity, 10) }],
-    };
-    // TODO replace for each mutation in the checkout thingy. can we export them from there???
-    // create your own custom hook???
-
-    lineItemAddMutation({ variables }).then((res) => {
-      setCartOpen(true);
-    });
-  };
-
-  const updateLineItemInCart = (lineItemId, quantity) => {
-    const variables = {
-      checkoutId: checkout.id,
-      lineItems: [{ id: lineItemId, quantity: parseInt(quantity, 10) }],
-    };
-    lineItemUpdateMutation({ variables });
-  };
-
-  const removeLineItemInCart = (lineItemId) => {
-    const variables = { checkoutId: checkout.id, lineItemIds: [lineItemId] };
-    lineItemRemoveMutation({ variables });
-  };
-
-  const associateCustomerCheckout = (customerAccessToken) => {
-    const variables = {
-      checkoutId: checkout.id,
-      customerAccessToken: customerAccessToken,
-    };
-    customerAssociateMutation({ variables }).then((res) => {
-      setCustomerAuthOpen(false);
-    });
-  };
+    shopLoading,
+    shopError,
+    closeCustomerAuth,
+    isCustomerAuthOpen,
+    isNewCustomer,
+    associateCustomerCheckout,
+    accountVerificationMessage,
+    shopData,
+    addVariantToCart,
+    checkout,
+  } = useContext(CartContext);
 
   if (shopLoading) {
     return <p>Loading ...</p>;
@@ -246,14 +45,7 @@ function Shop(props) {
           />
         ))}
       </div>
-      <Cart
-        removeLineItemInCart={removeLineItemInCart}
-        updateLineItemInCart={updateLineItemInCart}
-        checkout={checkout}
-        isCartOpen={isCartOpen}
-        handleCartClose={handleCartClose}
-        customerAccessToken={customerAccessToken}
-      />
+      {/* <Cart /> */}
     </div>
   );
 }
