--- src/contexts/OpenCartContext.js
@@ -1,12 +1,0 @@
-import React, { createContext, useState } from "react";
-
-export const OpenCartContext = createContext();
-
-export const OpenCartProvider = ({ children }) => {
-  const [isCartOpen, setCartOpen] = useState(false);
-  return (
-    <OpenCartContext.Provider value={{ isCartOpen, setCartOpen }}>
-      {children}
-    </OpenCartContext.Provider>
-  );
-};
